Fix to BED formatter allowing score to be floating point.
git-svn-id: https://svn2.assembla.com/svn/savant/Savant/trunk@1010 096565f4-0d05-4fcd-8cbf-ff6523290f6f

+++ src/savant/format/IntervalFormatter.java
@@ -251,7 +251,7 @@ public class IntervalFormatter extends SavantFileFormatter {
         fields.add(FieldType.INTEGER);  // start
         fields.add(FieldType.INTEGER);  // end
         fields.add(FieldType.STRING);   // name
-        fields.add(FieldType.INTEGER);  // score
+        fields.add(FieldType.FLOAT);  // score
         fields.add(FieldType.STRING);   // strand
         fields.add(FieldType.INTEGER);  // thickstart
         fields.add(FieldType.INTEGER);  // thickend
