+++ src/savant/format/SavantFileFormatterUtils.java
@@ -531,7 +531,7 @@ public class SavantFileFormatterUtils {
         int numFields = record.size();
         
         String name = "";
-        Integer score = 0;
+        Float score = 0.0f;
         Strand strand = SavantFileFormatterUtils.getStrand("+");
         Integer thickStart = 0;
         Integer thickEnd = 0;
@@ -539,7 +539,7 @@ public class SavantFileFormatterUtils {
         List<Block> blocks = null;
         
         if (numFields > 3) { name = (String) record.get(3); }
-        if (numFields > 4) { score = (Integer) record.get(4); }
+        if (numFields > 4) { score = (Float) record.get(4); }
         if (numFields > 5) { strand = SavantFileFormatterUtils.getStrand((String) record.get(5)); }
         if (numFields > 6) { thickStart = (Integer) record.get(6); } else { thickStart = 0; }
         if (numFields > 7) { thickEnd = (Integer) record.get(7); } else { thickEnd = interval.getLength(); }
